CTC Beam search - Maintain beamLength per beam and change beam history to not contain void symbols for padding
Summary:
It turned out that the 2 changes don't need to be done together, but I thought they were.
There is a small benefit from maintaining the number of actual symbols each beam will produce when its output is extracted.
We can reject any Merge attempts where the "copy candidate beam length + 1 != extend candidate beam length"
And output extraction is easier (But this only happens once).

The main change is to simplify Merge by simplifying the traceback mechanism (This too helps with output extraction but that only happens once).
The content of the beamParent data is changed to be an index into the flat beamparent array (It was a beam index, so implied that it referenced the previous timestep).
This means it can reference any timestep.
Then instead of inserting a void symbol we "patch up" the latest timestep to look like the previous one: The addend is the same addend as the previous one and the
parent is the parent from the previosu timestep.
All this means that when tracing back we have simpler code as we don't need to skip past any void symbols

Ref T35818

Test Plan: CI

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T35818

Differential Revision: https://phabricator.sourcevertex.net/D44184