Hide `{Unary,Binary,Ternary}OpNames` from the API
Summary:
Tensorflow poplar plugin often fails to load due to poplibs symbol errors.
One of the common errors comes from the declaration of the arrays that contain
the operations names, even though we don't need to expose them in the interface.

This diff moves the declaration and definition of the list of names away from
the public headers, and replaces vector and string types to use array and
string_view instead, which will allow the functions that access them to
directly access the read only section.

Test Plan: CI

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Maniphest Tasks: T38474

Differential Revision: https://phabricator.sourcevertex.net/D44397