Fix to edge case of empty beams outputs for CTC Inference
Summary:
In the case of a large beamwidth,  and fewer symbols than beams. If we get several  beam outputs, it is possible to return empty beam outputs when running for very few timesteps.  These should contain no data/have length =0.  This was wrong in both the model and the implementation.

Ref T35818

Test Plan: Specific short tests added

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T35818

Differential Revision: https://phabricator.sourcevertex.net/D44416