Cast to 8-bit integer should truncate
Summary:
Currently the popops cast of a floating point value to an 8-bit integer type
uses the rounding mode round-to-nearest-ties-to-even which is different from
casting to another integer type, which uses round-toward-zero (i.e. like the C++
function std::trunc()).
The latter is what is used in C++ static_cast and python casting from float to
integer, and so is probably the behaviour expected by users.

This changes the cast to 8-bit types to use round-toward-zero as well. The
penalty is a 33% speed decrease.

This change also reduces the code size: the two conversion loops do not need to
be RPT loops. They can use brnzdec and still be pipelined, with no additional
cycle overhead, and with the advantage of removing many nop instruction.

Also, the macro are slightly restructured so that it will become trivial to
reuse them to also implement a round/rint function in future

Test Plan: Run the existing tests, which have been modified to check for truncation instead of round-to-nearest

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Subscribers: yanislavd

Maniphest Tasks: T38209

Differential Revision: https://phabricator.sourcevertex.net/D44530