Removing limitations caused by number of tiles on CTC Inference
Summary:
Previously we needed tiles >= batchsize * (numSymbols-1)

This change removes that limitation, where tiles are split according to the batchsize (Each batch entry occupies an independent group of tiles if possible).  Then that number of tiles per batch entry is partitioned according to each functional step's requirements, attempting to partition as much as possible.
If the number of tiles < batchsize, partitioning will result in batch entries using overlapping tiles.

Eventually the only cause of failure ought to be running out of memory.

The fastest implementation will still be one that fragments the implementation over as many tiles/workers as possible.  This change will only affect what happens when that is not possible.
For a full IPU there will not be a limitation in a large number of cases.

This change is just to make those tile limited cases fit, not an attempt to optimise them for speed/ number of vertices created etc.
Ref T35818

Test Plan: Specific tests with limited tiles available

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Subscribers: samk

Maniphest Tasks: T35818

Differential Revision: https://phabricator.sourcevertex.net/D44460