Allow CTC inference loops to end early when all items in a batch are < maxT in length
Summary:
Like the title!

As a result of this we can make CTCUpdate a multivertex which speeds it up greatly.  The reason it couldn't be one before was that it updated the counter and when workers read the counter they could do so before/after it was updated by a specific worker

Ref T35818

Test Plan: CI

Reviewers: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T35818

Differential Revision: https://phabricator.sourcevertex.net/D44613