T22741: Remove dependency on poputil from poplibs_support
Summary:
The poplibs_support library uses some poputil exceptions but never
explicitly links to it in the build system. We get away with it most
of the time because the exceptions are defined in a header.

This patch replaces the poputil exceptions with poplar runtime errors.
This is technically a breaking API change but the exceptions are only
thrown for developer errors, so they should probably be fatal to the
application anyway.

Test Plan: CI.

Reviewers: saml, samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T22741

Differential Revision: https://phabricator.sourcevertex.net/D44737