CTC Inference rank vertex in assembler, with some renaming and less code duplication.
Summary:
This also absorbs the copy to initialise its outputs, and provides a total probability result, which is needed for multi-stage sort.

Rename SortSelect to SimpleSelect and remove sort and pre (of preSelect) prefixes where appropriate

More code duplication crept in, so added a function shared between codelet tests to cut down on that.  More of this is possible but getting even more off the main reason for this change!

Ref T31939

Test Plan: Added specific codelet test and also enabled the simpleSort codelet test

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T31939

Differential Revision: https://phabricator.sourcevertex.net/D44684