Reduce the need for copies in CTC Inference loop body
Summary:
Various changes to reduce the need for copies in the loop body of CTC inference.

The Extend candidates as an input to the select step require copying from the
tile they were made on onto the tile they needed for select.  By doing an
explicit copy at the same time as other copies the overhead for this is much reduced.

There is no need to make a copy of the beamLength data, instead use a ping-pong buffer
with previousValue, nextValue to manage the need for previousValue to be available.

previousLastBeamOutputs can be made redundant by using the last addend in the beam history
instead.  (lastBeamOutputs could probably also be eliminated?)

Ref T35818

Test Plan: CI

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T35818

Differential Revision: https://phabricator.sourcevertex.net/D44777