Fix void symbol comparison in CTC Loss codelets
Summary:
At time=0 beam histories are initialised with beam0=voidSymbol, beam1=invalidSymbol0, beam2=invalidSymbol1, ....
When the number of beams is greater than the number of symbols the invalidSymbols will propagate to the next timestep, as there are fewer candidates than beams - so the invalid beams and their "invalidSymbol" outputs survive the 1st timestep.

When taking a symbol from the beam history we were comparing it to the `voidSymbol`, and if not equal, the symbols probability was accessed using prob=logProbs[symbol].
This resulted in the logic if(symbol!=voidSymbol) { prob=logProbs[symbol]}  picking out invalidSymbols, and so accessing outside the bounds of the logProbs tensor data - and a segfault.
The condition needs to include the invalidSymbols as well as the voidSymbol

Note that in other places voidSymbol is directly inserted into the the addend of the copy candidate and so a comparison added==voidSymbol is correct.

Ref T35818

Test Plan: Specific test added to capture the initial failure case

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T35818

Differential Revision: https://phabricator.sourcevertex.net/D45057