Bug fix in beam history, plus using total probability to avoid logAdd in CTC Inference, and reduce, update vertex assembler
Summary:
Adding total probability to the beam history has very little cost and speeds up the candidate creation by avoiding logAdds in several places.

Making a multi-stage sort will require the Reduce step to cope with total probability as well, so that is added. (but not the multi stage sort yet)
Added assembler for Reduce vertex as it is simple and is one of the stages whose cost grows with large numbers of symbols
Added assembler for the Update vertex as it was taking a large proportion of the time for something so simple!
A bit more code sharing in the codelet test vertex generation

While developing this a bug cropped up, and the fix for that is included here:
We used to null out extend candidates that were merged just by zeroing the total probability.  PBlank, PNon-blank, addend and Parent were left as alone.
With a large beamwidth and a small number of symbols there can be fewer candidates (in early timesteps) than there are beams.  They would be candidates from invalid beams, sorted lower down the order but still within beamwidth from the top.
These had a zero totalProbability but a non-zero pBlank and pNonBlank.  That meant that when candidates were generated the resulting probability was not always zero (Could be derived from pBlank for example).
Also, the candidate has an addend, which was a valid symbol.  That meant that occasionally 2 beams (one that we want and the one from an extend candidate that should have been excluded) would apparently be the same.  This is an unintended state for beam history to be in.

To avoid the above, the addend was set to `voidSymbol` and where that is seen in the Update stage all 3  probabilities are set to zero.  Examining this made it clear that only a void symbol was needed, not a unique value for each invalid beam so that was simplified too.

Ref T35818

Test Plan: Add codelet test for Reduce vertex

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T35818

Differential Revision: https://phabricator.sourcevertex.net/D44859