Add a utility to pick up on striding in a set of intervals and use these to improve popops::reduce graph construction time.
Summary:
Ref T36684.

Reductions work with a particular 2-dimensional view of the input where outputs are the innermost dimension and the dimension to reduce is the outermost.
When we have a contiguous reduction (where the dimension to be reduced is actually the one contiguous in memory), when we query the tile mapping/contiguous regions these are all single elements with a regular stride between them.

This change finds striding in the intervals that form the contiguous regions and uses them to perform far fewer slices of the input tensor. This has a big impact because:
* Slicing itself is not that cheap, partially due to the effort to simplify where possible at the point of slicing - partially because each slice is a new tensor and tensor expression that needs allocating.
* The tensor formed by concatenating the slices is then connected to vertices and/or introspected to check if it is contiguous which requires iterating the regions which gets slower the more complex the expression is.

The end result is a big speedup in graph construction particularly where the input was contiguous in the dimension to be reduced.

e.g. an arbitrary reduce_op example I was using:
```
build/poplibs/tools/reduce_op --shape=2,10000000 --dims=1
```
this command was ~45s before, ~15s with D44173, and ~3s with this change also.

In combination with D44173:
 * Reduces BERT Large pre-training 384 seq size (Configuration from T35690) graph construction time from ~4 mins to ~2.5 mins.
 * Reduces "Structured BERT" reproducer (T36217) total Poplar compile time from ~33 mins to ~25mins.

Reviewers: richard, samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: timp

Maniphest Tasks: T36684

Differential Revision: https://phabricator.sourcevertex.net/D44247