CTC Inference Model faster Merge
Summary:
The merge stage was by far the slowest in the reference model
With 1866 symbols, beamwidth 5 and 50 timesteps it was taking more than 3 mins to run (I never waited long enough so it could be much much more)
After the change this became 15 seconds

Further speed improvements in merge would be had by trying to be clever about what we compare.  But for now attempting to keep it as dumb as possible to avoid introducing bugs.

Ref T35818

Test Plan: CI, Overlapping verification of new/old methods in place and kept for review reference until landing

Reviewers: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T35818

Differential Revision: https://phabricator.sourcevertex.net/D45523