Fix VectorInnerInPlace<MUL, half> scribbling with big input
Summary:
When the inner dimension is 4, the VectorInnerInPlace<MUL, half> vertex divides
the size of the outer dimension by 3 by multiplying it by the Q16.16 fixed
point approximation of 1/3, using the value 0x5556.
This gives correct results only if the value is less than 32768.
For greater values (up to 192602 =~3*2^16 = ) it gives approximated result,
which can be too big by 0, 1 or 2.
So for these values, corresponding to the size of the first operand >= 32767*4
(131072), the code will scribble past the end of the data.

The fix is to use the multiply in Q15.17 (i.e. use the value 0xaaab) which
gives correct results up to 98303, which correspond to a the size of the first
operand up to 98303 x 4 x 2 bytes = 786,424 bytes which is greater than the
maximum size of data that can be stored on current versions of the device.

Test Plan: manual test

Reviewers: Eddy, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: Eddy, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T39379

Differential Revision: https://phabricator.sourcevertex.net/D45679