Speed up cycle estimator for nx1 convolution vertex particularly for higher dimensional convolutions.
Summary:
Ref T33310.

We take advantage of the fact that we re-use the same worklists per kernel position in the planner to reduce the amount of work to estimate cycles in the nx1 convolution vertex. This helps particularly for large 3-dimensional convolutions where there the number of kernel elements is a large enough factor to cut down the estimation time considerably. This means a little extra code complexity (2 paths for kernel loop in nx1 cycle estimator) but tried to share as much code as possible.

Cuts down the following example from > 1 hour (didn't wait for it to finish) to ~6m30s:
```
/single_conv_layer --inference-only --batch-size 1 --field='{224,224,160}' --input-channels 32 --output-channels 32 --kernel-size='{3,3,3}' --data-type float --padding 1 --device-type IpuModel1
```

Reviewers: richard, samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: richard, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T33310

Differential Revision: https://phabricator.sourcevertex.net/D46286