Implement option to select float partials for pooling
Summary:
Provide internal calculation in float for accuracy, when in/out are half.
Only really relevant for sum, avg pool with half in, out

The main pooling implementation does this by casting at present, but doesn't have to be the case.
Reduction and convolution possibilities for implementing pooling can benefit (as casting isn't needed or happens internally as required), and (in a later change) grouping selections can also benefit as we know the output type now

The option may just be for integration with frameworks until supporting changes are made, so should remain internal for now maybe?
Because:
- Behaviour was to use half throughout - the option defaulting to half results in no change in behaviour by default.
- But (See task) tensorflow will always cast to float and back to improve accuracy. For accuracy, float should be selected similarly to reduction
- popart will not - half data will be used throughout
- Initially just providing the option will allow TF to not cast and therefore poplibs takes control of the implementation, which can then be improved (and the option possibly removed?) by only changing one repository.

Fixes T7322

Test Plan:
Added tests to use the option.
Nearly all tests pass with float partials as the default except a couple that go out of memory on Sim1 : Sim1_default_octconvHL_parallel,  Sim1_default_octconvHL_serial
Now the default is half partials so there is no issue

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T7322

Differential Revision: https://phabricator.sourcevertex.net/D46510