Remove redundant simple sort sorting method from CTC Inference
Summary:
We can't see a use case for this at the moment, and it only gets run when forced to in tests.

As it simplifies the plan variables to describe the sort step (By removing the need for a boost variant) also tidy those into a single vector of structs rather than multiple vectors.

Ref T35818

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml, oleksiik

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml, oleksiik

Maniphest Tasks: T35818

Differential Revision: https://phabricator.sourcevertex.net/D46662