Add the concept of user-facing 'priority group's to popsolver and use this in convolution planning.
Summary:
Ref T12879.

This is a 'major' priority which is considered before the size of a domain of a variable when choosing the next variable to try specific values of when minimising.
We form priority groups which are just groups of variables in the model, and define relative priorities between these groups. In the graph formed by these relative priorities, priority groups that are in sub-graphs that are not connected to one another are not prioritised over one another, and the depth of the priority group in each connected sub-graph is used to prioritise when priority groups are in the same connected sub-graph.

Note this also removes the implicit prioritisation of variables that are inputs to a call. This was prioritising inputs to calls that did not particularly cut the search space, and preventing us from evaluating variables that had much smaller domains that were more useful for cutting the search space.

Test Plan:
For the command
```
./build/poplibs/tools/single_conv_layer --inference-only --batch-size 1 --field='{224,224,160}' --input-channels 32 --output-channels 32 --kernel-size='{3,3,3}' --data-type float --padding 1 --device-type IpuModel1
```
with D46286 applied, this brings planning time down from ~6m30s to ~3m30s. It also reduces total number of constraints evaluated during propagation steps in the solver from just over 2546 million to 949 million.

Reviewers: richard, samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: richard, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T12879

Differential Revision: https://phabricator.sourcevertex.net/D46470