Add support for multi-update MAX as part of general support for scatter reduce operations
Summary:
Add assembler variants of codelets. A new API
MultiUpdateMax similar to MultiUpdateAdd is added
without a scale parameter. Extension to other
reduction operators is easily possible as the vertex
is templated with the operator enumeration type that
is the same as the existing operations type used in
reductions.

Ref T41378

Test Plan: New tests added

Reviewers: Eddy, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: Eddy, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T41378

Differential Revision: https://phabricator.sourcevertex.net/D47408