Add ability to plan minimising cycles to the embedding planner.
Summary:
Ref T40999. Fixes T12935.

Adds a way to plan an embedding layer to minimize cycles rather than our current weighted memory target. This currently defaults to the previous behaviour.

`planMinimisationTarget` currently controls the behaviour, and is intended to be a temporary option until a switch can be made to the new behaviour.
`availableMemoryProportion` now actually sets a hard limit on estimated temporary memory for a planned slice/update.

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, godfrey.da.costa

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, godfrey.da.costa

Subscribers: godfrey.da.costa

Maniphest Tasks: T40999, T12935

Differential Revision: https://phabricator.sourcevertex.net/D48101