Remove changes to tile linearisation order introduced by D48101.
Summary:
Ref T43684.

This was causing some new rearrangements in the test cases that regressed, and was never intended to be included in D48101 as it was an experiment.
I've otherwise maintained small fixes that I included in the original diff as they appear to have no effect on the plans produced in the regressions.

Also remove DynamicSlicePlanningTest. It's brittle and its purpose is served by regression tests which demonstrate the quality of the plan. This was originally to test how the planner behaved in the days before the implementation was done.

Reviewers: godfrey.da.costa, samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: godfrey.da.costa, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T43684

Differential Revision: https://phabricator.sourcevertex.net/D49343