Add a function to regroup a tensor if possible to rearrangement library
Summary:
Regrouping a tensor differs from regroupIfBeneficial
as there is no reference tensor to compare against.
Istead the tensor is introspected and the requested
regrouping performance is performed if possible.
The original tensor is returned if it is not possible
to regroup.

Ref T43807

Test Plan: New test added

Reviewers: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T43807

Differential Revision: https://phabricator.sourcevertex.net/D49445