Remove tests using old-style replicated graphs from poplibs.
Summary:
Ref T43915. Depends on D49659.

This removes the distributed batch norm test as there is no way to test this from poplibs without the old replication API (because we need gcl collectives/it is convenient to use gcl collectives to test single-image replicated distributed batch norm). A test in gcl to replace it is a dependency of this.

Reviewers: samk, godfrey.da.costa, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: godfrey.da.costa, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T43915

Differential Revision: https://phabricator.sourcevertex.net/D49660