Correction to expression names using unaryOpTypeToString etc
Summary:
Some of the unary ops were associated with the wrong name in the codelet name produced, but the code was functionally correct.
With just an array of strings it is easy to get out of sync with the actual names when any change is made, which is maybe what happened.
So using unary,binary,ternary opTypeToString is more maintainable.  The names are more verbose though - is this OK?  I prefer correct and verbose but if they are too long we could duplicate the unaryOpTypeToString with unaryOpTypeToShortString

Fixes T44420

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, samk

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T44420

Differential Revision: https://phabricator.sourcevertex.net/D49854