Allow independent planning of multi-update and multi-slice
Summary:
The existing implementation allows only multi-slice to be
planned independently as a forward pass of an embedding layer.
multi-update is always planned as a joint plan with multi-slice.
As gather and scatter operations can occur as operations in a
forward pass, we allow them to be planned independently.

An option `usedForSlice' which controls whether cost of
mult-slice is added to the planner.

Ref T44637

Reviewers: samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Subscribers: hatemh

Maniphest Tasks: T44637

Differential Revision: https://phabricator.sourcevertex.net/D50194