Opt out of aligning supervisor vertices
Summary:
Ref T17150.

A new pass was added to automatically align workers when stochastic rounding instructions are used by that codelet. The ability to opt-out was also added, this change marks all supervisor vertices as not needing alignment (because they use the worker id to index into the data, therefore it doesn't matter which worker starts first).

Reviewers: godfrey.da.costa, timp, tomm, oleksiik, claudiop, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: oleksiik, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T17150, T44044

Differential Revision: https://phabricator.sourcevertex.net/D50228