Pooling changes to ensure that all of the output tensor is initialised
Summary:
Backward pool can produce an output that has elements set to zero if the stride is larger than the kernel size.
The data in these locations wasn't always written to in the poplibs pool function. It can happen on tiles assigned part of the output which is all zero, or on a last fragment on a tile after the last non-zero value.

Additionally with input padding the same can happen in the fwd pass.

This isn't noticed in tests as data is initialised to zero.  It should be noticed in debug where data is initialised to nan, but I think because other operations (fwd pool) write to memory first it seems to be OK.  Making the test run bwd only ought to help to catch this problem, as will adding more tests.

Fixes T44777

Test Plan: Create bwd only test with stride > kernel size.

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, samk

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T44777

Differential Revision: https://phabricator.sourcevertex.net/D47557