Start tile for intermediate reductions should always be dithered
Summary:
The existing implementation specialises the case when there is
only one split map in the intermediate stages and starts adding
reductions from the first tile (as per the comment in the code
this is to avoid overloading tile 0). With dithering already
applied this should not happen anyway and we should just map
the reductions based on the start tile given by the dithering.

Ref T45046

Reviewers: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Maniphest Tasks: T45046

Differential Revision: https://phabricator.sourcevertex.net/D50572