Support planning of multiUpdate (scatter operation)
Summary:
Allows specifying a plan for a multi-update. This
requires an operation to be specified as part of
the options when planning. This brings it in line
with other gather/scatter operations that can be
planned. Look-up splits for multi-update are not
allowed as a reduction would be required and we
cannot in general define a reduction operator for
it. The planner therefore disallows a partition
of the look-up.

Ref T15083

Test Plan: New tests added

Reviewers: samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: hatemh

Maniphest Tasks: T15083

Differential Revision: https://phabricator.sourcevertex.net/D50678