Allow setting Boost_USE_STATIC* independently of BUILD_SHARED_LIBS
Summary:
I mistook BUILD_SHARED_LIBS for a good proxy for whether boost was built
statically or dynamically. It works for poplarview and colossus_tools_view, but
not in host_runtime_view as that view uses a mix of BUILD_SHARED_LIBS=ON/OFF in
its SuperProjectConfig.cmake file. (I tried updating host_runtime_view to build
only shared, but the buildbots failed it and I think this approach is better.)

This is a prerequisite for D49324.

Ref T43516, D49193, D49324.

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jorgec, samk

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Maniphest Tasks: T43516

Differential Revision: https://phabricator.sourcevertex.net/D50879