Allow scale in MultiUpdateAdd to be in float precision
Summary:
Allowing the scale to be in float precision makes
multiUpdateAdd consistent with other weight updates
for the case of embeddings and sparse updates of
weights.

There is no run-time check on precision of the scale
to decide on using float/half scale primarily
because the innermost loop costs only one cycle extra
compared to the case when half scale is used. We
can choose to use the run-time check if there are
use-cases that benefit from doing so.

Ref T36632

Test Plan: New tests added

Reviewers: Eddy, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: Eddy, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: oskarb

Maniphest Tasks: T36632

Differential Revision: https://phabricator.sourcevertex.net/D51079