Make convolution reduction stages respect convolution dithering - including output allocation
Summary:
Allocation used the smallest numbered tiles available, often resulting in tile 0 upwards being allocated.  Other candidate tiles are just as appropriate for the reduction operation in question (As they also contain partials) - so there are many equally valid choices.

Dithering enabled:
This change makes the allocation a true reflection of the dithered result - that is the start tile and direction of
allocation is respected.

Dithering disabled:
This change selects the last allocated tile to begin implementing reductions on, and therefor will select tiles below it for the remaining work.

Fixes T42430

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, oleksiik, tomm

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oleksiik

Maniphest Tasks: T42430

Differential Revision: https://phabricator.sourcevertex.net/D50995