Support multi-updates with sorted indices
Summary:
The information on whether indices are sorted or
not is provided by an option flag. Only a planned
update operation will target vertices that use
the sorted indices. The current vertices have
a field indicating sorting should be done or
not. A binary search is done per worker for the
indices to be processed.

This functionality should allow us to perform
a sorted segment sum.

Ref T39919

Test Plan: Codelets tests and planner tests added

Reviewers: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T39919

Differential Revision: https://phabricator.sourcevertex.net/D51480