Rearrange scaledAddSupervisor vertex state to save space, use 16 byte aligned scales
Summary:
There was a 16 bit hole in the vertex state - making the scale pointers into 16 bit pointers by using SCALED_PTR128 this can be packed in.  It also makes the state for the tensor version smaller than that of the constant version assuming that the same scale is used multiple times

All versions have vertex state rearranged to be consistent but the benefit is where we have a single Tensor scale.  This is the common case, and it is used in place of the single Constant scale cases in the vertex generating function

Ref T46026

Test Plan: CI

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, claudiop, oleksiik, tomm

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, claudiop

Maniphest Tasks: T46026

Differential Revision: https://phabricator.sourcevertex.net/D51803